Implement custom exceptions (fix: #3)

+++ src/main/java/org/trendwa/eshop/basketservice/service/RedisBasketService.java
@@ -2,6 +2,7 @@ package org.trendwa.eshop.basketservice.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.trendwa.eshop.basketservice.exception.CustomerBasketNotFoundException;
 import org.trendwa.eshop.basketservice.model.CustomerBasket;
 import org.trendwa.eshop.basketservice.repository.BasketRepository;
 
@@ -9,13 +10,17 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class BasketServiceImpl implements BasketService {
+public class RedisBasketService implements BasketService {
 
     private final BasketRepository basketRepository;
 
     @Override
-    public Optional<CustomerBasket> getBasketByBuyerId(String buyerId) {
-        return basketRepository.findById(buyerId);
+    public CustomerBasket getBasketByBuyerId(String buyerId) throws CustomerBasketNotFoundException {
+        Optional<CustomerBasket> basket = basketRepository.findById(buyerId);
+        if (basket.isEmpty()) {
+            throw new CustomerBasketNotFoundException(buyerId);
+        }
+        return basket.get();
     }
 
     @Override
@@ -25,6 +30,7 @@ public class BasketServiceImpl implements BasketService {
 
     @Override
     public void deleteBasketByBuyerId(String buyerId) {
-        basketRepository.deleteById(buyerId);
+        if(basketRepository.existsById(buyerId)) basketRepository.deleteById(buyerId);
+        throw new CustomerBasketNotFoundException(buyerId);
     }
 }
